licensing: restructure user limit enforcement file
