+++ enterprise/cmd/frontend/internal/licensing/enforcement/enforcement.go
@@ -1,56 +1,10 @@
-package licensing
+package enforcement
 
 import (
-	"context"
 	"fmt"
 	"html"
 	"net/http"
-
-	"github.com/inconshreveable/log15"
-
-	"github.com/sourcegraph/sourcegraph/internal/errcode"
 )
-
-// NewPreCreateUserHook returns a PreCreateUserHook closure with
-// the given UsersStore.
-func NewPreCreateUserHook(s UsersStore) func(context.Context) error {
-	return func(ctx context.Context) error {
-		info, err := GetConfiguredProductLicenseInfo()
-		if err != nil {
-			return err
-		}
-		var licensedUserCount int32
-		if info != nil {
-			licensedUserCount = int32(info.UserCount)
-		} else {
-			licensedUserCount = NoLicenseMaximumAllowedUserCount
-		}
-
-		// Block creation of a new user beyond the licensed user count (unless true-up is allowed).
-		userCount, err := s.Count(ctx)
-		if err != nil {
-			return err
-		}
-		// Be conservative and treat 0 as unlimited. We don't plan to intentionally generate
-		// licenses with UserCount == 0, but that might result from a bug in license decoding, and
-		// we don't want that to immediately disable Sourcegraph instances.
-		if licensedUserCount > 0 && int32(userCount) >= licensedUserCount {
-			if info != nil && info.HasTag(TrueUpUserCountTag) {
-				log15.Info("Licensed user count exceeded, but license supports true-up and will not block creation of new user. The new user will be retroactively charged for in the next billing period. Contact sales@sourcegraph.com for help.", "activeUserCount", userCount, "licensedUserCount", licensedUserCount)
-			} else {
-				message := "Unable to create user account: "
-				if info == nil {
-					message += fmt.Sprintf("a Sourcegraph subscription is required to exceed %d users (this instance now has %d users). Contact Sourcegraph to learn more at https://about.sourcegraph.com/contact/sales.", NoLicenseMaximumAllowedUserCount, userCount)
-				} else {
-					message += "the Sourcegraph subscription's maximum user count has been reached. A site admin must upgrade the Sourcegraph subscription to allow for more users. Contact Sourcegraph at https://about.sourcegraph.com/contact/sales."
-				}
-				return errcode.NewPresentationError(message)
-			}
-		}
-
-		return nil
-	}
-}
 
 // WriteSubscriptionErrorResponseForFeature is a wrapper around WriteSubscriptionErrorResponse that
 // generates the error title and message indicating that the current license does not active the
