+++ enterprise/cmd/frontend/internal/licensing/init/init.go
@@ -23,7 +23,7 @@ import (
 func Init(ctx context.Context, enterpriseServices *enterprise.Services) error {
 	// Enforce the license's max user count by preventing the creation of new users when the max is
 	// reached.
-	db.Users.PreCreateUser = licensing.NewPreCreateUserHook(&usersStore{})
+	db.Users.PreCreateUser = enforcement.NewPreCreateUserHook(&usersStore{})
 
 	// Enforce the license's max external service count by preventing the creation of new external
 	// services when the max is reached.
